fixed decimal point issue

+++ src/main/java/com/kalsym/deliveryservice/provider/PriceResult.java
@@ -7,6 +7,7 @@ package com.kalsym.deliveryservice.provider;
 import com.fasterxml.jackson.annotation.JsonInclude;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -16,7 +17,7 @@ import java.util.Date;
 public class PriceResult {
     public Long refId;
     public int providerId;
-    public double price;
+    public BigDecimal price;
     public String validUpTo;
     public String message;
     public boolean isError;
